refactor(svelte): split r0-evidence interaction suite by concern
Break the 1084-line R0 evidence monolith into tests/r0-evidence/
(fixtures + pointer-inspect / touch-interval / tooltip-zoom) so review
and failure localization match one responsibility per file. Assertion
bodies and it-titles are preserved; helpers share rows/size/pointer.
